Rewrite aoAll to work with Infinite time DNF
DNF can be treated like an infinite time
The mean of a set of times that includes DNF is DNF
The worst time in a set of times that includes DNF is DNF
A sorted set of times will have the DNFs at the end

This makes the DNFsort method obsolete.
We can also extract the getMeanTimeSeconds method
This method now works even if DNF times are included
Thus it can be used in other places.