Add balancedRandom util
Returns a "balanced" random index given counts.

Basically if you randomly pick a number between 1 and 5  1,000,000 times
you can expect with a high degree of certainty
that each number will show up about 200,000 times

However if you do the same thing, but only 10 times
it is highly unlikely that each number will appear exactly 2 times

Put simply, the distribution is only even for a large number of trials.

When testing case solves,
the user will typically select 2-10 cases
if they only do 20 solves,
there's a good chance that some cases will never be present
and others will be presented a lot of times

balanced random index fixes this problem by
- making each case appear at least once before any appear twice
- making cases that have appeared a lot less likely to appear

We cannot make it completely random,
or the distribution will be uneven for a small number of trials
And we cannot make it completely balanced,
or the user will be able to predict cases at the end of each cycle