Make Navbar Collapse on select on small screens
Grab your coffee!
1. A navbar normally allows the user to navigate via hyperlinks
When a user clicks a link, they are redirected to a new page
This effectively reloads the website
2. Bootstrap's navbar is just this sort of navbar.
it lets the user navigate via hyperlinks
3. In order to be responsive to small screens,
the bootstrap navbar collects the navigation links into a drop down
This happens when you access most websites using a phone
When an item on this drop down is selected, a new page is loaded
This new page has the dropdown collaped by default...
as such, the expanded dropdown is effectively collapsed
4. React router does not work using hyperlinks and doesn't reload pages.
Instead it uses the current url to conditionally render components
React router links are then used only to set the url to certain values
5. Since the page is not reloaded, bootstraps navbar does not collapse
It thinks you have not selected an item yet
6. React Bootstrap offers the collapseOnSelect property.
When this is set, clicking navlinks collapses the dropdown
However!, this only works if the nav links have event keys
I do not know why this is.
7 Event keys make nested nav links weird
If you add an event key to a navdropdown item it makes it behave weridly
In this particular exampe, adding a navdropdown to the settings link
Doing this makes the link and the other navlinks misbehave
More specifically, some links remain highlighted even when not...
on that page

8. The Solution
The react-bootstrap navbar has an...
expanded property and an onToggle property
onToggle is triggered when the dropdown icon is tapped
expanded controls whether the dropdown is expanded or not

using these two properties we can turn the navbar into...
a controlled component of sorts.
We lift the expanded state out of the navbar out where we can alter it
The collapse method can then be called anytime any link is clicked
This manually collapses the dropdown.

All this just to make the darn thing collapse on my phone
smh